part of the refactor
