+++ ElementX/Sources/Other/Pills/PillContext.swift
@@ -28,7 +28,7 @@ final class PillContext: ObservableObject {
     
     private var cancellable: AnyCancellable?
     
-    init(roomContext: RoomScreenViewModel.Context, data: PillTextAttachmentData) {
+    init(roomContext: TimelineViewModel.Context, data: PillTextAttachmentData) {
         switch data.type {
         case let .user(id):
             let isOwnMention = id == roomContext.viewState.ownUserID
@@ -73,7 +73,7 @@ extension PillContext {
         switch type {
         case .loadUser(let isOwn):
             pillType = .user(userID: testID)
-            let viewModel = PillContext(roomContext: RoomScreenViewModel.mock.context, data: PillTextAttachmentData(type: pillType, font: .preferredFont(forTextStyle: .body)))
+            let viewModel = PillContext(roomContext: TimelineViewModel.mock.context, data: PillTextAttachmentData(type: pillType, font: .preferredFont(forTextStyle: .body)))
             viewModel.viewState = PillViewState(isOwnMention: isOwn, displayText: testID)
             Task {
                 try? await Task.sleep(for: .seconds(2))
@@ -82,12 +82,12 @@ extension PillContext {
             return viewModel
         case .loadedUser(let isOwn):
             pillType = .user(userID: "@test:test.com")
-            let viewModel = PillContext(roomContext: RoomScreenViewModel.mock.context, data: PillTextAttachmentData(type: pillType, font: .preferredFont(forTextStyle: .body)))
+            let viewModel = PillContext(roomContext: TimelineViewModel.mock.context, data: PillTextAttachmentData(type: pillType, font: .preferredFont(forTextStyle: .body)))
             viewModel.viewState = PillViewState(isOwnMention: isOwn, displayText: "@Very Very Long Test Display Text")
             return viewModel
         case .allUsers:
             pillType = .allUsers
-            return PillContext(roomContext: RoomScreenViewModel.mock.context, data: PillTextAttachmentData(type: pillType, font: .preferredFont(forTextStyle: .body)))
+            return PillContext(roomContext: TimelineViewModel.mock.context, data: PillTextAttachmentData(type: pillType, font: .preferredFont(forTextStyle: .body)))
         }
     }
 }
