+++ ElementX/Sources/Screens/RoomScreen/RoomScreenCoordinator.swift
@@ -49,7 +49,8 @@ enum RoomScreenCoordinatorAction {
 }
 
 final class RoomScreenCoordinator: CoordinatorProtocol {
-    private var viewModel: RoomScreenViewModelProtocol
+    private var roomViewModel: RoomScreenViewModelProtocol
+    private var timelineViewModel: TimelineViewModelProtocol
     private var composerViewModel: ComposerToolbarViewModel
     private var wysiwygViewModel: WysiwygComposerViewModel
 
@@ -61,17 +62,20 @@ final class RoomScreenCoordinator: CoordinatorProtocol {
     }
     
     init(parameters: RoomScreenCoordinatorParameters) {
-        let viewModel = RoomScreenViewModel(roomProxy: parameters.roomProxy,
-                                            focussedEventID: parameters.focussedEventID,
-                                            timelineController: parameters.timelineController,
-                                            mediaProvider: parameters.mediaProvider,
-                                            mediaPlayerProvider: parameters.mediaPlayerProvider,
-                                            voiceMessageMediaManager: parameters.voiceMessageMediaManager,
-                                            userIndicatorController: ServiceLocator.shared.userIndicatorController,
-                                            appMediator: parameters.appMediator,
-                                            appSettings: parameters.appSettings,
-                                            analyticsService: ServiceLocator.shared.analytics)
-        self.viewModel = viewModel
+        let roomViewModel = RoomScreenViewModel()
+        self.roomViewModel = roomViewModel
+        
+        let timelineViewModel = TimelineViewModel(roomProxy: parameters.roomProxy,
+                                                  focussedEventID: parameters.focussedEventID,
+                                                  timelineController: parameters.timelineController,
+                                                  mediaProvider: parameters.mediaProvider,
+                                                  mediaPlayerProvider: parameters.mediaPlayerProvider,
+                                                  voiceMessageMediaManager: parameters.voiceMessageMediaManager,
+                                                  userIndicatorController: ServiceLocator.shared.userIndicatorController,
+                                                  appMediator: parameters.appMediator,
+                                                  appSettings: parameters.appSettings,
+                                                  analyticsService: ServiceLocator.shared.analytics)
+        self.timelineViewModel = timelineViewModel
 
         wysiwygViewModel = WysiwygComposerViewModel(minHeight: ComposerConstant.minHeight,
                                                     maxCompressedHeight: ComposerConstant.maxHeight,
@@ -80,12 +84,12 @@ final class RoomScreenCoordinator: CoordinatorProtocol {
         composerViewModel = ComposerToolbarViewModel(wysiwygViewModel: wysiwygViewModel,
                                                      completionSuggestionService: parameters.completionSuggestionService,
                                                      mediaProvider: parameters.mediaProvider,
-                                                     mentionDisplayHelper: ComposerMentionDisplayHelper(roomContext: viewModel.context),
+                                                     mentionDisplayHelper: ComposerMentionDisplayHelper(roomContext: timelineViewModel.context),
                                                      analyticsService: ServiceLocator.shared.analytics,
                                                      composerDraftService: parameters.composerDraftService)
         
         NotificationCenter.default.publisher(for: UIApplication.willResignActiveNotification).sink { _ in
-            viewModel.saveDraft()
+            roomViewModel.saveDraft()
         }
         .store(in: &cancellables)
     }
@@ -93,7 +97,7 @@ final class RoomScreenCoordinator: CoordinatorProtocol {
     // MARK: - Public
     
     func start() {
-        viewModel.actions
+        timelineViewModel.actions
             .sink { [weak self] action in
                 guard let self else { return }
 
@@ -123,7 +127,7 @@ final class RoomScreenCoordinator: CoordinatorProtocol {
                 case .displayLocation(let body, let geoURI, let description):
                     actionsSubject.send(.presentLocationViewer(body: body, geoURI: geoURI, description: description))
                 case .composer(let action):
-                    composerViewModel.process(roomAction: action)
+                    composerViewModel.process(timelineAction: action)
                 case .displayCallScreen:
                     actionsSubject.send(.presentCallScreen)
                 case .displayPinnedEventsTimeline:
@@ -136,21 +140,32 @@ final class RoomScreenCoordinator: CoordinatorProtocol {
             .sink { [weak self] action in
                 guard let self else { return }
 
-                viewModel.process(composerAction: action)
+                timelineViewModel.process(composerAction: action)
+            }
+            .store(in: &cancellables)
+        
+        roomViewModel.actions
+            .sink { [weak self] action in
+                guard let self else { return }
+
+                switch action {
+                case .composer(let action):
+                    composerViewModel.process(roomAction: action)
+                }
             }
             .store(in: &cancellables)
         
         // Loading the draft requires the subscriptions to be set up first otherwise the room won't be be able to propagate the information to the composer.
-        viewModel.loadDraft()
+        roomViewModel.loadDraft()
     }
     
     func focusOnEvent(eventID: String) {
-        Task { await viewModel.focusOnEvent(eventID: eventID) }
+        Task { await timelineViewModel.focusOnEvent(eventID: eventID) }
     }
     
     func stop() {
-        viewModel.saveDraft()
-        viewModel.stop()
+        roomViewModel.saveDraft()
+        timelineViewModel.stop()
     }
     
     func toPresentable() -> AnyView {
@@ -158,10 +173,12 @@ final class RoomScreenCoordinator: CoordinatorProtocol {
                                               wysiwygViewModel: wysiwygViewModel,
                                               keyCommands: composerViewModel.keyCommands)
 
-        return AnyView(RoomScreen(context: viewModel.context, composerToolbar: composerToolbar)
-            .onDisappear { [weak self] in
-                self?.viewModel.saveDraft()
-            })
+        return AnyView(RoomScreen(roomViewModel: roomViewModel,
+                                  timelineViewModel: timelineViewModel,
+                                  composerToolbar: composerToolbar)
+                .onDisappear { [weak self] in
+                    self?.roomViewModel.saveDraft()
+                })
     }
 }
