+++ ElementX/Sources/Screens/RoomScreen/View/Style/TimelineStyler.swift
@@ -62,7 +62,7 @@ struct TimelineStyler<Content: View>: View {
 }
 
 struct TimelineItemStyler_Previews: PreviewProvider, TestablePreview {
-    static let viewModel = RoomScreenViewModel.mock
+    static let viewModel = TimelineViewModel.mock
 
     static let base = TextRoomTimelineItem(id: .random,
                                            timestamp: "Now",
