+++ ElementX/Sources/Screens/RoomScreen/View/Timeline/AudioRoomTimelineView.swift
@@ -36,7 +36,7 @@ struct AudioRoomTimelineView: View {
 }
 
 struct AudioRoomTimelineView_Previews: PreviewProvider, TestablePreview {
-    static let viewModel = RoomScreenViewModel.mock
+    static let viewModel = TimelineViewModel.mock
     
     static var previews: some View {
         body.environmentObject(viewModel.context)
