+++ ElementX/Sources/Screens/RoomScreen/View/Timeline/CallInviteRoomTimelineView.swift
@@ -32,7 +32,7 @@ struct CallInviteRoomTimelineView: View {
 }
 
 struct CallInviteRoomTimelineView_Previews: PreviewProvider, TestablePreview {
-    static let viewModel = RoomScreenViewModel.mock
+    static let viewModel = TimelineViewModel.mock
     
     static var previews: some View {
         body.environmentObject(viewModel.context)
