+++ ElementX/Sources/Screens/RoomScreen/View/Timeline/CallNotificationRoomTimelineView.swift
@@ -19,7 +19,7 @@ import Foundation
 import SwiftUI
 
 struct CallNotificationRoomTimelineView: View {
-    @EnvironmentObject private var context: RoomScreenViewModel.Context
+    @EnvironmentObject private var context: TimelineViewModel.Context
     
     let timelineItem: CallNotificationRoomTimelineItem
     
@@ -62,7 +62,7 @@ struct CallNotificationRoomTimelineView: View {
 }
 
 struct CallNotificationRoomTimelineView_Previews: PreviewProvider, TestablePreview {
-    static let viewModel = RoomScreenViewModel.mock
+    static let viewModel = TimelineViewModel.mock
     
     static var previews: some View {
         body.environmentObject(viewModel.context)
