+++ ElementX/Sources/Screens/RoomScreen/View/Timeline/NoticeRoomTimelineView.swift
@@ -44,7 +44,7 @@ struct NoticeRoomTimelineView: View, TextBasedRoomTimelineViewProtocol {
 }
 
 struct NoticeRoomTimelineView_Previews: PreviewProvider, TestablePreview {
-    static let viewModel = RoomScreenViewModel.mock
+    static let viewModel = TimelineViewModel.mock
     
     static var previews: some View {
         body.environmentObject(viewModel.context)
