+++ ElementX/Sources/Screens/RoomScreen/View/Timeline/UnsupportedRoomTimelineView.swift
@@ -40,7 +40,7 @@ struct UnsupportedRoomTimelineView: View {
 }
 
 struct UnsupportedRoomTimelineView_Previews: PreviewProvider, TestablePreview {
-    static let viewModel = RoomScreenViewModel.mock
+    static let viewModel = TimelineViewModel.mock
     
     static var previews: some View {
         body.environmentObject(viewModel.context)
